requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
